+++ miso.cabal
@@ -154,6 +154,7 @@ library
     Miso.Event.Decoder
     Miso.Event.Types
     Miso.Exception
+    Miso.Fetch
     Miso.FFI
     Miso.FFI.History
     Miso.FFI.SSE
